chore: address issues with smoke tests
