--- cypress/support/step_definitions/annual-review/scheduler/eligible_list_items_are_correct.js
@@ -1,12 +1,0 @@
-Then("eligible list items are correct", async () => {
-  cy.task("db", {
-    operation: "list.findFirst",
-    variables: {
-      where: {
-        reference: "SMOKE",
-      },
-    },
-  }).then((list) => {
-    cy.expect(list.jsonData.currentAnnualReview.eligibleListItems.length).to.equal(6);
-  });
-});
