+++ cypress/support/step_definitions/annual-review/scheduler/there_are_no_eligible_list_items.js
@@ -1,0 +1,13 @@
+Then("there are no eligible list items", async () => {
+    cy.task("db", {
+      operation: "list.findFirst",
+      variables: {
+        where: {
+          reference: "SMOKE",
+        },
+      },
+    }).then((list) => {
+        cy.log(list.jsonData)
+      cy.expect(list.jsonData).to.not.have.property('currentAnnualReview');
+    });
+  });
