+++ cypress/support/step_definitions/annual-review/scheduler/there_are_string_eligible_list_items.js
@@ -1,0 +1,12 @@
+Then("there are {string} eligible list items", async (listItem) => {
+  cy.task("db", {
+    operation: "list.findFirst",
+    variables: {
+      where: {
+        reference: "SMOKE",
+      },
+    },
+  }).then((list) => {
+    cy.expect(list.jsonData.currentAnnualReview.eligibleListItems.length).to.equal(Number(listItem));
+  });
+});
