+++ cypress/support/step_definitions/annual-review/scheduler/there_is_an_eligible_list_item.js
@@ -1,0 +1,12 @@
+Then("there is an eligible list item", async () => {
+    cy.task("db", {
+      operation: "list.findFirst",
+      variables: {
+        where: {
+          reference: "SMOKE",
+        },
+      },
+    }).then((list) => {
+      cy.expect(list.jsonData.currentAnnualReview.eligibleListItems.length).to.equal(1);
+    });
+  });
